cambios orientados a la mejora visual y tecnica del codigo
