+++ pagina/src/app.css
@@ -9,6 +9,7 @@
     margin: auto;
     top: 0;
     left: 0;
+    text-align: center;
 }
 
 body {
