+++ pagina/src/app.html
@@ -1,5 +1,6 @@
 <!doctype html>
 <html lang="es">
+	<title>AUDM</title>
 	<head>
 		<meta charset="utf-8" />
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
