+++ pagina/src/app2.css
@@ -58,6 +58,7 @@ nav{
     max-width: 900px;
     margin: auto;
     padding-bottom: 50px;
+    text-align: center;
 }
 
 a{color: rgb(0, 60, 255)}
