+++ pagina/src/routes/+page.server.ts
@@ -1,8 +1,8 @@
-import { lucia } from "$lib/database/lucia";
 import { redirect } from "@sveltejs/kit";
 export const load = (async ({locals}) => {
-    if (!locals.user){
-        return redirect(302,'/inicio')
+    if (!locals.user)
+    {
+        return redirect(302,'/registro')
     }
-    return redirect(302,'/pagina/inicio')
+    return redirect(302,'/paginas/inicio')
 });
