--- pagina/src/routes/pagina/+layout.svelte
@@ -1,15 +1,0 @@
-<script>
-    import Cabezo from './cabezo.svelte'
-    import Pie from './pie.svelte'
-    import '../../app.css'
-    import '../../app2.css'
-</script>
-
-<Cabezo/>
-
-<main>
-    <slot/>
-</main>
-
-<Pie/>
-
