--- pagina/src/routes/pagina/cabezo.svelte
@@ -1,30 +1,0 @@
-<header>
-    <nav>
-        <ul class="menu-horizontal">
-            <li>
-                <a href="/pagina/inicio"
-                    ><img
-                        src="https://res.cloudinary.com/dnontz2gc/image/upload/v1723574291/img%20pia/house.png"
-                        alt="casa"
-                        width="50px"
-                        height="50px"
-                    /></a
-                >
-            </li>
-            <li><a href="/pagina/inventario" class="inv">Inventario</a></li>
-            <li>
-                <a href="/pagina/acerca-de" class="inv">Información</a>
-            </li>
-            <li>
-                <a href="../../../inicio_out"
-                    ><img
-                        src="https://cdn-icons-png.flaticon.com/512/61/61208.png"
-                        alt="pato"
-                        width="50px"
-                        height="50px"
-                    /></a
-                >
-            </li>
-        </ul>
-    </nav>
-</header>
