--- pagina/src/routes/pagina/inventario/apartados/+layout.svelte
@@ -1,19 +1,0 @@
-<html lang="es">
-    <title>Inventario AUDM</title>
-</html>
-<main>
-    <div class="texto_1">
-        <center>
-            <a href="../"
-                ><img
-                    src="https://cdn-icons-png.flaticon.com/512/61/61208.png"
-                    alt="pato"
-                    width="50px"
-                    height="50px"
-                /></a
-            >
-
-            <slot />
-        </center>
-    </div>
-</main>
