+++ pagina/src/routes/paginas/+layout.server.ts
@@ -1,0 +1,10 @@
+import {redirect } from "@sveltejs/kit";
+
+export const load = (async ({locals}) => {
+    if (!locals.user) 
+	{
+		return redirect(302, "/inicio");
+	}
+	return {user: locals.user};
+});
+
