+++ pagina/src/routes/paginas/+layout.svelte
@@ -1,0 +1,11 @@
+<script>
+    import '../../app.css';
+    import '../../app2.css';
+    import Estructura from '../../components/estructura.svelte';
+</script>
+
+<Estructura>
+    <main>
+        <slot/>
+    </main>
+</Estructura>
