+++ pagina/src/routes/paginas/inventario/apartados/+layout.svelte
@@ -1,0 +1,13 @@
+<main>
+    <div class="texto_1">
+        <a href="../">
+            <img
+                src="https://cdn-icons-png.flaticon.com/512/61/61208.png"
+                alt="pato"
+                width="50px"
+                height="50px"
+            />
+        </a>
+        <center><slot/></center>
+    </div>
+</main>
